tutorials/How_to_automate_Notion_with_Python_Notion_API_Tutorial/pyproject.toml
[project]
name = "how-to-automate-notion-with-python-notion-api-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
